requirements.txt
openai
ultralytics
transformers
torch
Pillow
language-tool-python
textstat
flask
flask-cors